% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_h5}
\alias{load_h5}
\title{Load an hdf5 file produced by the Fisher Stepping Stone model.}
\usage{
load_h5(file, dataset = NULL)
}
\arguments{
\item{file}{Path to the file.}

\item{dataset}{optional specific dataset to open}
}
\value{
A list containing all datasets and parameters from the file.
Or a simple array when a specific dataset is chosen.
}
\description{
Load an hdf5 file produced by the Fisher Stepping Stone model,
either entirelly or only for a specified dataset.
}
\examples{
file <- system.file("extdata", "result_sim.h5", package = "rfss")
result_sim <- load_h5(file)

}
