% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_animation}
\alias{plot_animation}
\title{Produce an animation combining all generations}
\usage{
plot_animation(res, outdir = "./")
}
\arguments{
\item{res}{A dataset produced by \code{\link{load_h5}}.}

\item{outdir}{Where to save the html file.}
}
\value{
html file, images/ folder, css/ folder and js/ folder.
}
\description{
Produce an animation combining all generations as an html file.
}
\examples{
data(result_sim)
plot_animation(result_sim)

}
