% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_timestep}
\alias{plot_timestep}
\title{Combined plot of mean fitness and allele frequencies for each deme}
\usage{
plot_timestep(data, gen = 1, message = T)
}
\arguments{
\item{data}{A dataset produced by \code{\link{load_h5}}.}

\item{gen}{The generation at which you whish to plot (default 1).}

\item{message}{Option to remove the message indicating the closest real
generation available that has been plotted (default TRUE).}
}
\value{
A plot generated by \code{\link{gridExtra::grid.arrange}}
}
\description{
Combined plot of mean fitness and allele frequencies for each deme
}
\examples{
data(result_sim)
plot_timestep(result_sim, 11)
plot_timestep(result_sim, 10) # produces the same plot

}
